+++ infra/tf/s3_minio/main.tf
@@ -34,6 +34,6 @@ provider "aws" {
 	s3_use_path_style = true
 	endpoints {
 		s3 = local.s3_endpoint_internal
-	}	
+	}
 }