+++ lib/bolt/config/src/ns.rs
@@ -212,24 +212,30 @@ pub struct CloudflareAccessServices {
 pub struct S3 {
 	#[serde(default)]
 	pub cors: S3Cors,
+	pub backfill: Option<String>,
 	#[serde(flatten)]
-	pub provider: S3Provider,
+	pub providers: S3Providers,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug, Default)]
 #[serde(deny_unknown_fields)]
 pub struct S3Cors {
-	#[serde(default)]
 	pub allowed_origins: Option<Vec<String>>,
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, Default)]
+#[serde(deny_unknown_fields)]
+pub struct S3Providers {
+	pub minio: Option<S3Provider>,
+	pub backblaze: Option<S3Provider>,
+	pub aws: Option<S3Provider>,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]
 #[serde(deny_unknown_fields)]
-pub enum S3Provider {
-	#[serde(rename = "minio")]
-	Minio {},
-	#[serde(rename = "backblaze")]
-	Backblaze {},
+pub struct S3Provider {
+	#[serde(default)]
+	pub default: bool,
 }
 
 #[derive(Serialize, Deserialize, Clone, Debug)]