+++ lib/bolt/core/src/context/service.rs
@@ -15,7 +15,7 @@ use crate::{
 		self,
 		service::{RuntimeKind, ServiceKind},
 	},
-	context::{self, BuildContext, ProjectContext, RunContext},
+	context::{self, BuildContext, ProjectContext, RunContext, S3Provider},
 	dep::{self, cloudflare, s3},
 	utils,
 };
@@ -297,6 +297,10 @@ impl ServiceContextData {
 	pub fn depends_on_s3(&self) -> bool {
 		true
 		// self.name().starts_with("upload-")
+	}
+
+	pub fn depends_on_s3_backfill(&self) -> bool {
+		self.name() == "upload-provider-fill"
 	}
 
 	pub fn depends_on_cloudflare(&self) -> bool {
@@ -975,7 +979,7 @@ impl ServiceContextData {
 			env.push((format!("CRDB_URL_{}", crdb_dep.name_screaming_snake()), uri));
 		}
 
-		// Expose all S3 endpoints to services that need it
+		// Expose all S3 endpoints to services that need them
 		let s3_deps = if self.depends_on_s3() {
 			project_ctx.all_services().await.to_vec()
 		} else {
@@ -986,35 +990,56 @@ impl ServiceContextData {
 				continue;
 			}
 
-			let s3_config = project_ctx
-				.clone()
-				.s3_config(project_ctx.clone().s3_credentials().await?)
-				.await?;
+			// Add default provider
+			let default_provider_name = match project_ctx.default_s3_provider()? {
+				(S3Provider::Minio, _) => "MINIO",
+				(S3Provider::Backblaze, _) => "BACKBLAZE",
+				(S3Provider::Aws, _) => "AWS",
+			};
+			env.push((
+				"S3_DEFAULT_PROVIDER".to_string(),
+				default_provider_name.to_string(),
+			));
 
-			env.push((
-				format!("S3_BUCKET_{}", s3_dep.name_screaming_snake()),
-				s3_dep.s3_bucket_name().await,
-			));
-			env.push((
-				format!("S3_ENDPOINT_INTERNAL_{}", s3_dep.name_screaming_snake()),
-				s3_config.endpoint_internal,
-			));
-			env.push((
-				format!("S3_ENDPOINT_EXTERNAL_{}", s3_dep.name_screaming_snake()),
-				s3_config.endpoint_external,
-			));
-			env.push((
-				format!("S3_REGION_{}", s3_dep.name_screaming_snake()),
-				s3_config.region,
-			));
-			env.push((
-				format!("S3_ACCESS_KEY_ID_{}", s3_dep.name_screaming_snake()),
-				s3_config.credentials.access_key_id,
-			));
-			env.push((
-				format!("S3_SECRET_ACCESS_KEY_{}", s3_dep.name_screaming_snake()),
-				s3_config.credentials.access_key_secret,
-			));
+			// Add all configured providers
+			let providers = &project_ctx.ns().s3.providers;
+			if providers.minio.is_some() {
+				add_s3_env(
+					&project_ctx,
+					&mut env,
+					&s3_dep,
+					context::project::S3Provider::Minio,
+				)
+				.await?;
+			}
+			if providers.backblaze.is_some() {
+				add_s3_env(
+					&project_ctx,
+					&mut env,
+					&s3_dep,
+					context::project::S3Provider::Backblaze,
+				)
+				.await?;
+			}
+			if providers.aws.is_some() {
+				add_s3_env(
+					&project_ctx,
+					&mut env,
+					&s3_dep,
+					context::project::S3Provider::Aws,
+				)
+				.await?;
+			}
+		}
+
+		// S3 backfill
+		if self.depends_on_s3_backfill() {
+			if let Some(backfill) = &project_ctx.ns().s3.backfill {
+				env.push((
+					"S3_BACKFILL_PROVIDER".into(),
+					heck::ShoutySnakeCase::to_shouty_snake_case(backfill.as_str()),
+				));
+			}
 		}
 
 		// Runtime-specific
@@ -1225,4 +1250,47 @@ impl ServiceContextData {
 
 		service
 	}
+}
+
+async fn add_s3_env(
+	project_ctx: &ProjectContext,
+	env: &mut Vec<(String, String)>,
+	s3_dep: &Arc<ServiceContextData>,
+	provider: S3Provider,
+) -> Result<()> {
+	let provider_name = match provider {
+		S3Provider::Minio => "MINIO",
+		S3Provider::Backblaze => "BACKBLAZE",
+		S3Provider::Aws => "AWS",
+	};
+	let s3_dep_name = s3_dep.name_screaming_snake();
+	let s3_config = project_ctx.s3_config(provider).await?;
+	let s3_creds = project_ctx.s3_credentials(provider).await?;
+
+	env.push((
+		format!("S3_{}_BUCKET_{}", provider_name, s3_dep_name),
+		s3_dep.s3_bucket_name().await,
+	));
+	env.push((
+		format!("S3_{}_ENDPOINT_INTERNAL_{}", provider_name, s3_dep_name),
+		s3_config.endpoint_internal,
+	));
+	env.push((
+		format!("S3_{}_ENDPOINT_EXTERNAL_{}", provider_name, s3_dep_name),
+		s3_config.endpoint_external,
+	));
+	env.push((
+		format!("S3_{}_REGION_{}", provider_name, s3_dep_name),
+		s3_config.region,
+	));
+	env.push((
+		format!("S3_{}_ACCESS_KEY_ID_{}", provider_name, s3_dep_name),
+		s3_creds.access_key_id,
+	));
+	env.push((
+		format!("S3_{}_SECRET_ACCESS_KEY_{}", provider_name, s3_dep_name),
+		s3_creds.access_key_secret,
+	));
+
+	Ok(())
 }