+++ svc/pkg/upload/ops/complete/src/lib.rs
@@ -1,8 +1,15 @@
 use futures_util::stream::{StreamExt, TryStreamExt};
-use proto::backend::pkg::*;
+use proto::backend::{self, pkg::*};
 use rivet_operation::prelude::*;
 use serde_json::json;
 use std::{collections::HashMap, time::Duration};
+
+#[derive(Debug, sqlx::FromRow)]
+struct UploadRow {
+	bucket: String,
+	user_id: Option<Uuid>,
+	provider: i64,
+}
 
 #[derive(Debug, sqlx::FromRow)]
 struct FileRow {
@@ -20,14 +27,14 @@ async fn handle(
 
 	let upload_id = internal_unwrap!(ctx.upload_id).as_uuid();
 
-	let (bucket, files, user_id) = fetch_files(&crdb, upload_id).await?;
+	let (bucket, provider, files, user_id) = fetch_files(&crdb, upload_id).await?;
 	let files_len = files.len();
 
 	if let Some(req_bucket) = &ctx.bucket {
 		assert_eq_with!(&bucket, req_bucket, DB_INVALID_BUCKET);
 	}
 
-	let s3_client = s3_util::Client::from_env(&bucket).await?;
+	let s3_client = s3_util::Client::from_env_with_provider(&bucket, provider).await?;
 
 	let nsfw_scores =
 		validate_profanity_scores(&ctx, &s3_client, upload_id, &files, user_id).await?;
@@ -82,32 +89,42 @@ async fn handle(
 async fn fetch_files(
 	crdb: &CrdbPool,
 	upload_id: Uuid,
-) -> GlobalResult<(String, Vec<FileRow>, Option<Uuid>)> {
-	let (bucket, user_id) = sqlx::query_as::<_, (String, Option<Uuid>)>(indoc!(
-		"
-		SELECT bucket, user_id
-		FROM uploads
-		WHERE upload_id = $1
-		"
-	))
-	.bind(upload_id)
-	.fetch_one(crdb)
-	.await?;
+) -> GlobalResult<(String, s3_util::Provider, Vec<FileRow>, Option<Uuid>)> {
+	let (upload, files) = tokio::try_join!(
+		sqlx::query_as::<_, UploadRow>(indoc!(
+			"
+			SELECT bucket, provider, user_id
+			FROM uploads
+			WHERE upload_id = $1
+			"
+		))
+		.bind(upload_id)
+		.fetch_one(crdb),
+		sqlx::query_as::<_, FileRow>(indoc!(
+			"
+			SELECT path, content_length, nsfw_score_threshold, multipart_upload_id
+			FROM upload_files
+			WHERE upload_id = $1
+			"
+		))
+		.bind(upload_id)
+		.fetch_all(crdb)
+	)?;
 
-	let files = sqlx::query_as::<_, FileRow>(indoc!(
-		"
-		SELECT path, content_length, nsfw_score_threshold, multipart_upload_id
-		FROM upload_files
-		WHERE upload_id = $1
-		"
-	))
-	.bind(upload_id)
-	.fetch_all(crdb)
-	.await?;
+	// Parse provider
+	let proto_provider = internal_unwrap_owned!(
+		backend::upload::Provider::from_i32(upload.provider as i32),
+		"invalid upload provider"
+	);
+	let provider = match proto_provider {
+		backend::upload::Provider::Minio => s3_util::Provider::Minio,
+		backend::upload::Provider::Backblaze => s3_util::Provider::Backblaze,
+		backend::upload::Provider::Aws => s3_util::Provider::Aws,
+	};
 
-	tracing::info!(?bucket, files_len = ?files.len(), "fetched files");
+	tracing::info!(bucket=?upload.bucket, ?provider, files_len = ?files.len(), "fetched files");
 
-	Ok((bucket, files, user_id))
+	Ok((upload.bucket, provider, files, upload.user_id))
 }
 
 async fn validate_profanity_scores(