+++ svc/pkg/upload/ops/get/src/lib.rs
@@ -10,6 +10,7 @@ struct UploadRow {
 	complete_ts: Option<i64>,
 	deleted_ts: Option<i64>,
 	user_id: Option<Uuid>,
+	provider: i64,
 }
 
 impl From<UploadRow> for backend::upload::Upload {
@@ -22,6 +23,7 @@ impl From<UploadRow> for backend::upload::Upload {
 			complete_ts: val.complete_ts,
 			deleted_ts: val.deleted_ts,
 			user_id: val.user_id.map(Into::into),
+			provider: val.provider as i32,
 		}
 	}
 }
@@ -47,7 +49,8 @@ async fn handle(
 			content_length,
 			complete_ts,
 			deleted_ts,
-			user_id
+			user_id,
+			provider
 		FROM uploads
 		WHERE upload_id = ANY($1)
 		"