+++ svc/pkg/upload/ops/prepare/src/lib.rs
@@ -30,9 +30,30 @@ async fn handle(
 	ctx: OperationContext<upload::prepare::Request>,
 ) -> GlobalResult<upload::prepare::Response> {
 	let crdb = ctx.crdb("db-upload").await?;
+	let provider = if let Some(provider) = ctx.provider {
+		let proto_provider = internal_unwrap_owned!(
+			backend::upload::Provider::from_i32(provider),
+			"invalid upload provider"
+		);
+
+		match proto_provider {
+			backend::upload::Provider::Minio => s3_util::Provider::Minio,
+			backend::upload::Provider::Backblaze => s3_util::Provider::Backblaze,
+			backend::upload::Provider::Aws => s3_util::Provider::Aws,
+		}
+	} else {
+		s3_util::Provider::default()?
+	};
+	// Convert back to proto to include the default
+	let proto_provider = match provider {
+		s3_util::Provider::Minio => backend::upload::Provider::Minio,
+		s3_util::Provider::Backblaze => backend::upload::Provider::Backblaze,
+		s3_util::Provider::Aws => backend::upload::Provider::Aws,
+	};
 
 	let s3_client_external =
-		s3_util::Client::from_env_opt(&ctx.bucket, s3_util::EndpointKind::External).await?;
+		s3_util::Client::from_env_opt(&ctx.bucket, provider, s3_util::EndpointKind::External)
+			.await?;
 
 	// Validate upload sizes
 	let total_content_length = ctx
@@ -96,14 +117,14 @@ async fn handle(
 		"
 		WITH
 			_insert_upload AS (
-				INSERT INTO uploads (upload_id, create_ts, content_length, bucket, user_id)
-				VALUES ($1, $2, $3, $4, $5)
+				INSERT INTO uploads (upload_id, create_ts, content_length, bucket, user_id, provider)
+				VALUES ($1, $2, $3, $4, $5, $6)
 				RETURNING 1
 			),
 			_insert_files AS (
 				INSERT INTO upload_files (upload_id, path, mime, content_length, nsfw_score_threshold)
 				SELECT $1, rows.*
-				FROM unnest($6, $7, $8, $9) AS rows
+				FROM unnest($7, $8, $9, $10) AS rows
 				RETURNING 1
 			)
 		SELECT 1
@@ -115,6 +136,7 @@ async fn handle(
 	.bind(total_content_length as i64)
 	.bind(&ctx.bucket)
 	.bind(user_id)
+	.bind(proto_provider as i32 as i64)
 	// Files
 	.bind(paths)
 	.bind(mimes)