+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 -   **Bolt** Support for connecting to Redis databases with `bolt redis sh`
+-   **Infra** Support configuring multiple S3 providers
+-   **Infra** Support multipart uploads
 -   **Infra** Replace Promtail-based log shipping with native Loki Docker driver
 -   **Infra** Add local Traefik Cloudflare proxy daemon for connecting to Cloudflare Access services
 
@@ -23,5 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+-   **Portal** Skip captcha if no Turnstile key provided
 -   **Infra** Resolve [RUSTSEC-2023-0044](https://rustsec.org/advisories/RUSTSEC-2023-0044)
--   Skip captcha if no Turnstile key provided
+-   **Bolt** Correctly resolve project root when building services natively